requirements.txt
pip                 22.3.1
Django              4.1.5
djangorestframework 3.14.0
exceptiongroup      1.1.0
pip                 22.3.1
